% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_between.R
\name{distance_between}
\alias{distance_between}
\title{Calculate Distance in Meters between Location and Nearest Stations}
\usage{
distance_between(stations, location, n_max = 10L)
}
\arguments{
\item{stations}{Selection from \code{dwd_stations()}.}

\item{location}{Result of \code{address_to_lonlat}.}

\item{n_max}{Number of entries to be returned. Default ist set to 10.}
}
\value{
Data frame containing list of stations in neighbourhood of specified location. From nearest to longest.
}
\description{
Calculate Distance in Meters between Location and Nearest Stations
}
\examples{
distance_between(stations = dwd_stations(), location = address_to_lonlat("Zugspitze, Deutschland"))
}
