% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Wrapper to download data}
\usage{
download_data(job, cleanup = TRUE)
}
\arguments{
\item{job}{Possible truncated result of \code{zip_urls()} containing urls and base names of station ids which should be downloaded.}

\item{cleanup}{Should the download folder be deleted after the process of unzipping, reading, and merging is completed?}
}
\value{
A data frame.
}
\description{
Wrapper to download data
}
