% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_stations.R
\name{dwd_stations}
\alias{dwd_stations}
\title{Donwload list of all DWD stations}
\usage{
dwd_stations(
  url_txt =
    "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/recent/KL_Tageswerte_Beschreibung_Stationen.txt"
)
}
\arguments{
\item{url_txt}{Default exists. Do not touch. The .txt-file can be found under \href{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/recent/KL_Tageswerte_Beschreibung_Stationen.txt}{DWD station list}.}
}
\value{
A data frame.
}
\description{
Donwload list of all DWD stations
}
\note{
Link and list contain German words.
}
\examples{
dwd_stations()
}
