% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_urls.R
\name{zip_urls}
\alias{zip_urls}
\title{Return urls of zip files if station data is available}
\usage{
zip_urls(link)
}
\arguments{
\item{link}{API link, e.g. \href{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/air_temperature/recent/}{here}}
}
\value{
A data frame containing station ids, basenames, and zip urls.
}
\description{
Return urls of zip files if station data is available
}
\examples{
zip_urls(link = "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/air_temperature/recent/")
}
